+++ src/dungeon/utility/Direction.java
@@ -1,4 +1,4 @@
-package dungeon;
+package dungeon.utility;
 
 public enum Direction {
     UP, DOWN, LEFT, RIGHT
